refactor auth actions and user data actions
